test(server): keep runtime host tests in approved homes
